PageToOdtPlugin:
Initial import.
This simplistic version generates a valid odt file,
where the headers (= =, == ==, etc) are mapped to the
'Titre 1' to 'Titre 4' (it's in french for now, but
will be configurable in a next version).
The other lines are put in "Standard" style.
For now, one may consider this more as a proof of
concept than an actual plugin.